Won't pass tests: updating the ETL topic mapping code
For xpro, prolearn, and mitxonline, this reconfigures the ETL pipelines to pull topic data from the mappings in the database rather than a CSV file. This requires some test changes that haven't happened yet.

This also adds a "full name" that generates a topic name with the parent topic in the manner that the code expects. This really should only be used here as it's recursive and can potentially be expensive to trigger.