When adjusting facets or the search term, fire a `search_update` event. Add PostHog dependency to ol-components.
In either case, this depends on a transformer being installed to convert the query string to the properties needed to do reporting on. The event fires without any additional data for facets, and only with a flag (isEnter) to determine how the user submitted their request when fired by the search term input.

This also adds PostHog as a requirement for the ol-components workspace - we need to know whether or not the user hit Enter or clicked the Submit icon, so we're capturing the PH event in the ol-components code.