Finishing this up; should work now (to the extent that it works)
- Removed the PostHogRoute component as it didn't really work consistently
- Added a EcommerceFeature component to do the feature flag checking in a standardized way
- Removed a bunch of stuff from the "Cart" Page (since it was just the Privacy Policy anyway)

Not using the PostHogFeature component here - I was trying, but I kept running into issues where it'd throw to the fallback if the flag was enabled. It seems much more stable now with the useFeatureFlagEnabled but I wouldn't be opposed to doing it manually, since this doesn't trigger the Feature Flag View event (and then we can also just.. wait for the flag to come back as something).