Reimplenting PostHog support for the NextJS version of the app
- Now pull the API key and host from `process.env`
- Now pull API host - we weren't doing this before; you probably don't want to specify this, but if you want to self-host Pos
tHog (it is an option!) you'd need to be able to specify this.
- Roll feature flag processing into the new Webpack Config
- Minor update to the LearningResourceDrawer to make the custom `lrd_view` hook there work

Going forward, things that use PostHog through the usePostHog hook (or one of their components) should work as expected. (The lrd_view one did some checking to make sure PostHog was turned on, hence it needed a bit more work.)